backend/requirements.txt
pycodestyle
flask
python-dotenv
flask-sqlalchemy
flask-migrate
flask-security-too
# wtforms
setuptools
argon2_cffi